# Remove final fields startIO and startEC